// src/app/modules/hcm/shared/models/holidays/holidays-configuration.ts
export class HolidaysConfiguration{
    idHolidaysConfiguration: number = -1;
    idCompany: number = -1;
    idTypeSalary: number = -1;
    typeSalary: string = '';
    idTypeDay: number = -1;
    typeDay: string = '';
    dayMax: number = -1;
    dayMin: number = -1;
    dayMinBonus: number = -1;
    minAdditionalDays: number = -1;
    maxAdditionalDays: number = -1;
    daysOld: number = -1;
    maxYear: number = -1;
    enableAnticipated: boolean = false;
    userCreate: string = "";
    userUpdate: string = "";
}